refactor: move the PVM2 cross-compile recipe into nub-build
The `riscv64emc-pvm2` target spec and the cross-compile flags that
produce a PVM2 guest are ISA facts, not JAVM facts — but they lived in
`build-javm`, so nub had no way to build a guest program even after
gaining `nub-linker`.

- `nub_build::pvm2::build_elf` owns the target JSON and the cross-compile
  (moved verbatim, including the `--inline-threshold=265` flag that every
  pinned gas vector depends on). `nub_build::pvm2::build` goes further and
  writes a linked `.nubp` ProgramBlob.
- The split exists so a personality can stop at the ELF and supply its own
  container: `build-javm` calls `build_elf`, then wraps and SSZ-encodes.
  It is down to the two lines that are actually about JAVM.
- `nub-build`'s existing x86 guest-kernel builder moves to an `arch_x86`
  module, re-exported as `nub_build::build` so `javm/build.rs` is
  untouched. The crate root now documents the two unrelated guest worlds
  it serves: PVM2 programs that run *inside* the engine, and the ring-0
  substrate that *hosts* it.

Guest blobs remain byte-identical: the target JSON moved unchanged and
lands at the same `$OUT_DIR/targets/` path either way.
