+++ nub/nub-build/Cargo.toml
@@ -3,7 +3,9 @@ name = "nub-build"
 version.workspace = true
 edition.workspace = true
 license.workspace = true
-description = "build.rs helper: cross-compile a nub bare-metal Arch guest crate (e.g. nub-arch-x86) for a bare-metal target"
+description = "build.rs helpers: cross-compile nub guest crates (PVM2 programs, bare-metal arch guest kernels)"
 
 [dependencies]
 build-crate = { path = "../../rust/build-crate" }
+nub-linker = { path = "../nub-linker" }
+nub-program = { path = "../nub-program" }
